revert(pypi): revert the default for pipstar (#3373)
A user mentioned that flipping the pipstar default caused
a regression in their setup in [this comment]. Until we have
a better understanding and a regression test for that, we should
revert the flip. Adjusted the CHANGELOG notes for this.

Related to #2949

[this comment]:
https://github.com/bazel-contrib/rules_python/issues/2949#issuecomment-3417501543

(cherry picked from commit 3c3e0be43fce2cdf279fa42252fc958c05eb0779)