tests: set --windows_enable_symlinks in bzlmod example (#3409)
The `//tests:version_test_binary_*` targets fail when run on Bazel 8
with Windows with a "permission denied" error with the Python
executable. The notable thing about these tests is they're shell tests
with a data dependency on the Python program, so a symlink is created to
refer to the Python program. As best I can tell, Bazel 8 isn't creating
the symlink quite right, so Windows considers the file non-executable.

Setting `--windows_enable_symlinks` fixes this, but it isn't clear why.
Presumably it's telling Bazel to create symlinks more properly.