refactor(gazelle): Generate a modules map per wheel, then merge (#3415)
This change internally splits modules mapping generation to be
per-wheel, with a final quick "merge" action at the end.

The idea is to make this process both concurrent and cached (courtesy of
Bazel), which can be ideal for codebases with a large set of
requirements (as many monorepos end up doing)

Note that the `generator.py` interface changed. This seemed internal, so
I didn't mark it breaking (but this change could actually just leave the
generator alone, since the current implementation is fine with 1 wheel).

I ran this on the work repo and saw no change in output (but as I edited
a single requirement, the overall process was fast ⚡ )