feat(toolchain): drop 3.8 and print info level messages about it (#3387)
Before this PR we had to have at least one 3.8 toolchain to not break
things.
With this we should be good to drop it.

Any python_version 3.8 registrations will be dropped if there are no
actual
URLs configured, which means that 3.8 will not be selected. The same
with
pip.parse, we will just ignore it and won't add it to the hub. 

In order to ensure that `is_python_3.x` flags continue working, we
just alias them to `@platforms//:incompatible`. No deprecation message
is
printed.

Work towards #2704 

Next step for anyone interested and who has more time than me these
days:
- [ ] Remove the 3.9 URLs and add them individually to our examples to
show
      that one can do that.
- [ ] Update the examples to no longer use 3.9, because it is a
maintenance burden.