chore(bazelrc): fix the update-deleted-packages script (#3425)
Before this we were relying on an upstream script, but because how we
are building our docs we need to customize how it is working. What is
more, the format was hard to maintain and prone to human errors.

This splits out the deleted packages code into a separate rc file that
is much easier to work with.

Split out of #3421