fix(pip): do not add a pip-fallback when there is no sdist (#3432)
Before this PR the user would have to specify `download = True` in order
to not
fallback to pip, which is admittedly an odd interface design. With this
PR we
correctly do not add a `pip` fallback if there is no `sdist` to be used.
With
this in place we are better placed to enable the
`experimental_index_url` by
default. What is more we can more reliably detect when we should use a
special
repository rule to prepare for building from sdist.

Summary:
- Add a test that shows the problem and then adjust the torch
experimental URL
  test.
- Cleanup unused marker stubs in the tests.
- Fix the code to better handle the case when there is no sdist.

Work towards #260
Work towards #2410