fix(pip): set better defaults for the new target_platforms attr (#3447)
Fixup the default values for the `target_platforms` so that the
users can actually switch it as per docs.

I have also taken the liberty to update all of the tests to better
reflect how we set things up for legacy and index-url setups. As
it is now, legacy `bzlmod` whl_library layout is much more similar
to WORKSPACE, which makes the transition easier.

Work towards #2949